Se han realizado cambios estéticos y seguridad en la eliminación de la página de gestión de razas (faltan unos detalles por implementar)

+++ src/app/pages/raza/raza.component.ts
@@ -66,10 +66,15 @@ export class RazaComponent implements OnInit {
   }
 
   eliminarRaza(id: number): void {
-    this.razaService.delete(id).subscribe(() => {
-      this.cargarRazas();
-    });
+    const confirmacion = window.confirm('¿Estás seguro de que deseas eliminar esta raza?');
+
+    if (confirmacion) {
+      this.razaService.delete(id).subscribe(() => {
+        this.cargarRazas();
+      });
+    }
   }
+
 
   cancelarEdicion(): void {
     this.resetForm();
